Playback --------
Access to the underlying Engine is not recommended and now hidden. All
synthesis is done at the Session level. For that, playEngine is removed
from the API. An Session can be customized via playSession, which takes
the args passed to create an Engine

All offline rendering code was moved to core/offline.py

Playback envelopes have been redesigned. Before, fadeout was implemented
as an extension of the duration of the event. This led to difficult
legato lines in instruments like piano. Now fadeout does not extend
the duration. For piano legato it is much easier to use the sustain
param, which now correctly controls the extension of an event for
playback only.

More default dynamic parameters: both panning position and gain are
dynamic parameters now, meaning that they can be automated/set
for any defined preset.

Buses now reflect the changes done in csoundengine and work reliably
both live and offline

Misc
-----

* redesigned aliases in playback parameters, following csoundengine
* simplified API exposed in maelzel.core
* add overtake to synth automation
* documentation to builtin presets